# packages/backend/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "0.1.0",
		"title": "WAFRN Backend",
		"description": "API routes for wafrn social network"
	},
	"paths": {
...
		"/api/block": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/unblock": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/myBlocks": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/unblock-user": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/follow": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/unfollow": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v2/post/{id}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
		"/api/register": {
			"post": {
				"description": "",
				"responses": {
					"200": {
						"description": "OK"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"url": {
										"example": "any"
									},
									"email": {
										"example": "any"
									},
									"description": {
										"example": "any"
									},
									"name": {
										"example": "any"
									},
									"nsfw": {
										"example": "any"
									},
									"password": {
										"example": "any"
									},
									"birthDate": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/login": {
			"post": {
				"description": "",
				"responses": {
					"200": {
						"description": "OK"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"password": {
										"example": "any"
									},
									"email": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v2/dashboard": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "level",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
		"/post/{id}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "digest",
						"in": "header",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "user-agent",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"429": {
						"description": "Too Many Requests"
					}
				}
			}
		},
		"/blog/{url}/{somethingElse}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "url",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "somethingElse",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/blog/{url}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "url",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}